pyproject.toml
# =============================================================================
# pyproject.toml
# Triplet-MHA ViT — Project Metadata and Dependency Specification
#
# Build backend : setuptools (PEP 517/518 compliant)
# Install       : pip install -e .         (editable/development mode)
#               : pip install .            (regular install)
# =============================================================================

[build-system]
requires      = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

# ── Project metadata ──────────────────────────────────────────────────────
[project]
name        = "triplet-mha-vit"
version     = "0.1.0"
description = "Triplet Multi-Head Attention Vision Transformer for Plant Disease Detection"
readme      = "README.md"
license     = { file = "LICENSE" }
authors     = [
    { name = "Md Ashraful Haque", email = "ashrafulhaque664@gmail.com" }
]
keywords    = [
    "vision-transformer", "plant-disease-detection", "multi-head-attention",
    "deep-learning", "tensorflow", "computer-vision", "agriculture"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"

# ── Core runtime dependencies ─────────────────────────────────────────────
dependencies = [
    "tensorflow>=2.13.0",
    "numpy>=1.24.0",
    "Pillow>=9.4.0",
    "matplotlib>=3.6.0",
    "scikit-learn>=1.2.0",
    "pandas>=1.5.0",
    "xlsxwriter>=3.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.64.0",
]

# ── Optional dependency groups ────────────────────────────────────────────
[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
tensorboard = [
    "tensorboard>=2.10.0",
]
notebook = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

# ── Console entry points ──────────────────────────────────────────────────
[project.scripts]
triplet-train    = "scripts.train:main"
triplet-evaluate = "scripts.evaluate:main"

# ── URLs ──────────────────────────────────────────────────────────────────
[project.urls]
Homepage   = "https://github.com/mah-iasri/triplet-mha-vit"
Repository = "https://github.com/mah-iasri/triplet-mha-vit"
Issues     = "https://github.com/mah-iasri/triplet-mha-vit/issues"

# ── Setuptools package discovery ──────────────────────────────────────────
[tool.setuptools.packages.find]
where   = ["."]
include = ["models*", "utils*", "scripts*", "configs*"]

# ── Black formatter ───────────────────────────────────────────────────────
[tool.black]
line-length    = 100
target-version = ["py38", "py39", "py310"]

# ── isort import sorter ───────────────────────────────────────────────────
[tool.isort]
profile     = "black"
line_length = 100

# ── Pytest configuration ──────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths    = ["tests"]
python_files = ["test_*.py"]
addopts      = "-v --tb=short"